UserStoryOverviewStyle.css
This is not my original styling sheet but styling is not the hard part. Getting this to work with Scene Builder is my issue. I reference this in the fxml file and I go into scene builder to assign it to the anchor page. I've spent a ton of time trying to figure this out so if anyone can help, I'd be appreciative.